refactor(iam-policy): pointer and nil pointer refactor
