+++ resources/iam-policy.go
@@ -1,0 +1,148 @@
+package resources
+
+import (
+	"context"
+
+	"github.com/sirupsen/logrus"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/iam"
+	"github.com/aws/aws-sdk-go/service/iam/iamiface"
+
+	"github.com/ekristen/libnuke/pkg/registry"
+	"github.com/ekristen/libnuke/pkg/resource"
+	"github.com/ekristen/libnuke/pkg/types"
+
+	"github.com/ekristen/aws-nuke/v3/pkg/nuke"
+)
+
+const IAMPolicyResource = "IAMPolicy"
+
+func init() {
+	registry.Register(&registry.Registration{
+		Name:   IAMPolicyResource,
+		Scope:  nuke.Account,
+		Lister: &IAMPolicyLister{},
+		DependsOn: []string{
+			IAMUserPolicyAttachmentResource,
+			IAMGroupPolicyAttachmentResource,
+			IAMRolePolicyAttachmentResource,
+		},
+		DeprecatedAliases: []string{
+			"IamPolicy",
+		},
+	})
+}
+
+type IAMPolicyLister struct{}
+
+func (l *IAMPolicyLister) List(_ context.Context, o interface{}) ([]resource.Resource, error) {
+	opts := o.(*nuke.ListerOpts)
+
+	svc := iam.New(opts.Session)
+
+	params := &iam.ListPoliciesInput{
+		Scope: aws.String("Local"),
+	}
+
+	policies := make([]*iam.Policy, 0)
+
+	if err := svc.ListPoliciesPages(params,
+		func(page *iam.ListPoliciesOutput, lastPage bool) bool {
+			for _, listedPolicy := range page.Policies {
+				policy, err := GetIAMPolicy(svc, listedPolicy.Arn)
+				if err != nil {
+					logrus.Errorf("Failed to get listed policy %s: %v", *listedPolicy.PolicyName, err)
+					continue
+				}
+				policies = append(policies, policy)
+			}
+			return true
+		}); err != nil {
+		return nil, err
+	}
+
+	resources := make([]resource.Resource, 0)
+
+	for _, out := range policies {
+		resources = append(resources, &IAMPolicy{
+			svc:      svc,
+			name:     *out.PolicyName,
+			path:     *out.Path,
+			arn:      *out.Arn,
+			policyID: *out.PolicyId,
+			tags:     out.Tags,
+		})
+	}
+
+	return resources, nil
+}
+
+type IAMPolicy struct {
+	svc      iamiface.IAMAPI
+	name     string
+	policyID string
+	arn      string
+	path     string
+	tags     []*iam.Tag
+}
+
+func (r *IAMPolicy) Remove(_ context.Context) error {
+	resp, err := r.svc.ListPolicyVersions(&iam.ListPolicyVersionsInput{
+		PolicyArn: &r.arn,
+	})
+	if err != nil {
+		return err
+	}
+
+	for _, version := range resp.Versions {
+		if !*version.IsDefaultVersion {
+			_, err = r.svc.DeletePolicyVersion(&iam.DeletePolicyVersionInput{
+				PolicyArn: &r.arn,
+				VersionId: version.VersionId,
+			})
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	_, err = r.svc.DeletePolicy(&iam.DeletePolicyInput{
+		PolicyArn: &r.arn,
+	})
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (r *IAMPolicy) Properties() types.Properties {
+	properties := types.NewProperties()
+
+	properties.Set("Name", r.name)
+	properties.Set("ARN", r.arn)
+	properties.Set("Path", r.path)
+	properties.Set("PolicyID", r.policyID)
+	for _, tag := range r.tags {
+		properties.SetTag(tag.Key, tag.Value)
+	}
+	return properties
+}
+
+func (r *IAMPolicy) String() string {
+	return r.arn
+}
+
+// -------------
+
+func GetIAMPolicy(svc *iam.IAM, policyArn *string) (*iam.Policy, error) {
+	resp, err := svc.GetPolicy(&iam.GetPolicyInput{
+		PolicyArn: policyArn,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return resp.Policy, nil
+}
